Minor changes to the viewmodel and the mainActivity

+++ app/src/main/java/com/android/sneakerdroid/Participant.kt
@@ -12,7 +12,7 @@ data class Participant (
     var phoneNumber: String,
     var projectCode: String,
     var appVersion: Int,
-    var fcmKey: Int,
+    var fcmKey: Long,
     var deviceDetails: DeviceDetails) : Parcelable
